+++ worker/include/Logger.hpp
@@ -472,9 +472,8 @@ public:
 	{ \
 		std::fprintf(stderr, "(ABORT) " _MS_LOG_STR_DESC desc _MS_LOG_SEPARATOR_CHAR_STD, _MS_LOG_ARG, ##__VA_ARGS__); \
 		std::fflush(stderr); \
-		char abortMessage[Logger::BufferSize]; \
-		std::snprintf(abortMessage, Logger::BufferSize, "(ABORT) " _MS_LOG_STR_DESC desc _MS_LOG_SEPARATOR_CHAR_STD, _MS_LOG_ARG, ##__VA_ARGS__); \
-		throw std::runtime_error(abortMessage); \
+		std::snprintf(Logger::buffer, Logger::BufferSize, "(ABORT) " _MS_LOG_STR_DESC desc _MS_LOG_SEPARATOR_CHAR_STD, _MS_LOG_ARG, ##__VA_ARGS__); \
+		throw std::runtime_error(Logger::buffer); \
 	} \
 	while (false)
 #endif