+++ rust/CHANGELOG.md
@@ -5,7 +5,7 @@
 ### 0.24.1
 
 - Worker: Fix, use `thread_local` buffer on `MS_ABORT()` ([PR#1873](https://github.com/versatica/mediasoup/pull/1873)).
-- Handle subchannels in pipe `DataConsumers` ([PR#XXXX](https://github.com/versatica/mediasoup/pull/XXXX)).
+- Handle subchannels in pipe `DataConsumers` ([PR #1875](https://github.com/versatica/mediasoup/pull/1875)).
 
 ### 0.24.0