+++ rust/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### NEXT
 
+### 0.24.2
+
 ### 0.24.1
 
 - Worker: Fix, use `thread_local` buffer on `MS_ABORT()` ([PR#1873](https://github.com/versatica/mediasoup/pull/1873)).